requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
docker-pycreds==0.4.0
filelock==3.18.0
flwr==1.18.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.71.0
hydra-core==1.3.2
idna==3.10
importlib_resources==6.5.2
iterators==0.0.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lightning==2.5.1.post0
lightning-utilities==0.14.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
networkx==3.4.2
nibabel==5.3.2
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.7
propcache==0.3.1
protobuf==4.25.7
psutil==7.0.0
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytorch-lightning==2.5.1.post0
pytz==2025.2
PyYAML==6.0.2
ray==2.31.0
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.24.0
sentry-sdk==2.27.0
setproctitle==1.3.6
setuptools==80.3.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sympy==1.14.0
tomli==2.2.1
tomli_w==1.2.0
torch==2.7.0
torchmetrics==1.7.1
tqdm==4.67.1
triton==3.3.0
typer==0.12.5
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wandb==0.19.10
yarl==1.20.0
